fix: resolve mypy type errors in cerelog_eeg.py, lsl_gateway.py, tier1_coordinator.py
- cerelog_eeg.py: Add Optional guards for _board_adapter, _manager, _start_time; fix duration_s calculation
- lsl_gateway.py: Fix numpy dtype from float32 to float64 for push_chunk and get_recent_data return type
- tier1_coordinator.py: Fix PodState type confusion between Tier0PodState/Tier1PodState, add Path import, fix channels.get() union with explicit int type annotation

+++ src/synapse24/acquisition/lsl_gateway.py
@@ -331,7 +331,7 @@ class LSLGateway:
                         continue
 
                     # Convert to numpy
-                    data = np.array(chunk, dtype=np.float32)  # (n_samples, n_channels)
+                    data = np.array(chunk, dtype=np.float64)  # (n_samples, n_channels)
                     ts = np.array(timestamps, dtype=np.float64)
 
                     # Apply clock correction
@@ -450,7 +450,7 @@ class LSLGateway:
         if pod_id not in self._buffers:
             return None
 
-        data = np.array(self._buffers[pod_id], dtype=np.float32)
+        data = np.array(self._buffers[pod_id], dtype=np.float64)
         timestamps = np.array(self._timestamps[pod_id], dtype=np.float64)
 
         if duration_s is not None and len(timestamps) > 0:
